chore(git): Relocate v2 DDL to make future diffs easier to track
